requirements.txt
crewai==1.10.1
crewai-tools==0.12.1
streamlit==1.37.1
pandas==2.2.2
requests==2.32.3
openpyxl==3.1.5
openrouteservice==2.3.3
python-dotenv==1.1.1